stemai-backend/app/config.py
from pydantic import BaseModel
from functools import lru_cache
from dotenv import load_dotenv
import os

load_dotenv(dotenv_path=os.path.join(os.path.dirname(os.path.dirname(__file__)), ".env"))

class Settings(BaseModel):
    MONGODB_URI: str = os.getenv("MONGODB_URI", "mongodb://localhost:27017")
    MONGODB_DB: str = os.getenv("MONGODB_DB", "stemai")
    APP_ENV: str = os.getenv("APP_ENV", "development")

@lru_cache(maxsize=1)
def get_settings() -> Settings:
    return Settings()
